Guild creation implemented

+++ src/player/player/guild/add_guild_creation_player.rs
@@ -12,7 +12,7 @@ use eolib::{
 
 use crate::SETTINGS;
 
-use super::Player;
+use super::super::Player;
 
 impl Player {
     pub async fn add_guild_creation_player(&mut self, player_id: i32, name: String) {
@@ -22,7 +22,7 @@ impl Player {
 
         self.guild_create_members.push(player_id);
 
-        let packet = if self.guild_create_members.len() + 1 >= SETTINGS.guild.min_players as usize {
+        let packet = if self.guild_create_members.len() + 1 >= SETTINGS.guild.min_players {
             GuildReplyServerPacket {
                 reply_code: GuildReply::CreateAddConfirm,
                 reply_code_data: Some(GuildReplyServerPacketReplyCodeData::CreateAddConfirm(
